Set a min job size when parsing, fixing DST
Parsing a single line is very fast (10K lines is parsed in ~1ms on my
system). At that level, the overhead/synchronization for stealing work
is significant and counter-productive.

Similarly, fixing the date's DST is ~5x faster the parsing

There is no gain when applying the same change in the date check,
presumably because the bottleneck is memory access, not the processing.
And that pass is already very fast compared to the others, so the
overall benefit wouldn't be significant anyway.